GG-33174: nulls from percent to number, increnent version of stored partition statistics, fix cost function.

+++ modules/indexing/src/main/java/org/apache/ignite/internal/processors/query/stat/view/StatisticsColumnLocalDataView.java
@@ -104,7 +104,7 @@ public class StatisticsColumnLocalDataView {
      * @return Percent of nulls values.
      */
     @Order(6)
-    public int nulls() {
+    public long nulls() {
         return statistics.columnStatistics(column).nulls();
     }
